Services/Analytics/requirements.txt
google-generativeai
python-dotenv
langchain
fastapi
uvicorn[standard]  
